requirements.txt
pandas
nltk
sklearn_crfsuite
pytest
matplotlib
seaborn
ipykernel
datasets==3.0.1
streamlit==1.32.0
transformers==4.45.1
torch==2.2.2
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
fastapi
uvicorn
pydantic
jupyter 
ipywidgets
accelerate


